Bluetooth: ISO: Pass BIG encryption info through QoS
Mainline: 1d11d70d1f6b23e7d3fc00396c17b90b876162a4
From: v6.7-rc1
Severity: Low

[ Upstream commit 1d11d70d1f6b23e7d3fc00396c17b90b876162a4 ]

This enables a broadcast sink to be informed if the PA
it has synced with is associated with an encrypted BIG,
by retrieving the socket QoS and checking the encryption
field.

After PA sync has been successfully established and the
first BIGInfo advertising report is received, a new hcon
is added and notified to the ISO layer. The ISO layer
sets the encryption field of the socket and hcon QoS
according to the encryption parameter of the BIGInfo
advertising report event.

After that, the userspace is woken up, and the QoS of the
new PA sync socket can be read, to inspect the encryption
field and follow up accordingly.

Signed-off-by: Iulia Tanasescu <iulia.tanasescu@nxp.com>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
Stable-dep-of: 181a42edddf5 ("Bluetooth: Make handle of hci_conn be unique")
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 2341747553ef1e45469a33864c91e1dab0a46c5a)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I3ef46e974fb55a3b5956275aba7ce6efdd57684b
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83647
Tested-by: k2ci <kernel-bot@kylinos.cn>