s390/ap: re-init AP queues on config on
Mainline: 32d1d9204f8db3360be55e65bd182a1a68f93308
From: v6.7-rc1
Severity: Low

[ Upstream commit 32d1d9204f8db3360be55e65bd182a1a68f93308 ]

On a state toggle from config off to config on and on the
state toggle from checkstop to not checkstop the queue's
internal states was set but the state machine was not
nudged. This did not care as on the first enqueue of a
request the state machine kick ran.

However, within an Secure Execution guest a queue is
only chosen by the scheduler when it has been bound.
But to bind a queue, it needs to run through the initial
states (reset, enable interrupts, ...). So this is like
a chicken-and-egg problem and the result was in fact
that a queue was unusable after a config off/on toggle.

With some slight rework of the handling of these states
now the new function _ap_queue_init_state() is called
which is the core of the ap_queue_init_state() function
but without locking handling. This has the benefit that
it can be called on all the places where a (re-)init
of the AP queue's state machine is needed.

Fixes: 2d72eaf036d2 ("s390/ap: implement SE AP bind, unbind and associate")
Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>
Reviewed-by: Holger Dengler <dengler@linux.ibm.com>
Signed-off-by: Vasily Gorbik <gor@linux.ibm.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit a03b8ab0291e3c14a2b21b3c6c9af29f143e76c5)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: If2cd38dc7e68f25386c0a3b48e012be59bd0aa45
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83962
Tested-by: k2ci <kernel-bot@kylinos.cn>