sched/uclamp: Ignore (util == 0) optimization in feec() when p_util_max = 0
Mainline: 23c9519def98ee0fa97ea5871535e9b136f522fc
From: v6.7-rc1
Severity: Low

[ Upstream commit 23c9519def98ee0fa97ea5871535e9b136f522fc ]

find_energy_efficient_cpu() bails out early if effective util of the
task is 0 as the delta at this point will be zero and there's nothing
for EAS to do. When uclamp is being used, this could lead to wrong
decisions when uclamp_max is set to 0. In this case the task is capped
to performance point 0, but it is actually running and consuming energy
and we can benefit from EAS energy calculations.

Rework the condition so that it bails out when both util and uclamp_min
are 0.

We can do that without needing to use uclamp_task_util(); remove it.

Fixes: d81304bc6193 ("sched/uclamp: Cater for uclamp in find_energy_efficient_cpu()'s early exit condition")
Signed-off-by: Qais Yousef (Google) <qyousef@layalina.io>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Reviewed-by: Vincent Guittot <vincent.guittot@linaro.org>
Reviewed-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lore.kernel.org/r/20230916232955.2099394-3-qyousef@layalina.io
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 294e3c797d7cb27415f3a842f27e263e1a851ddd)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I226fbe15c17a50a6d81c3b9a8c6444733f3cb87f
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83508
Tested-by: k2ci <kernel-bot@kylinos.cn>