leds: pwm: Don't disable the PWM when the LED should be off
Mainline: 76fe464c8e64e71b2e4af11edeef0e5d85eeb6aa
From: v6.7-rc1
Severity: Low

[ Upstream commit 76fe464c8e64e71b2e4af11edeef0e5d85eeb6aa ]

Disabling a PWM (i.e. calling pwm_apply_state with .enabled = false)
gives no guarantees what the PWM output does. It might freeze where it
currently is, or go in a High-Z state or drive the active or inactive
state, it might even continue to toggle.

To ensure that the LED gets really disabled, don't disable the PWM even
when .duty_cycle is zero.

This fixes disabling a leds-pwm LED on i.MX28. The PWM on this SoC is
one of those that freezes its output on disable, so if you disable an
LED that is full on, it stays on. If you disable a LED with half
brightness it goes off in 50% of the cases and full on in the other 50%.

Fixes: 41c42ff5dbe2 ("leds: simple driver for pwm driven LEDs")
Reported-by: Rogan Dawes <rogan@dawes.za.net>
Reported-by: Fabio Estevam <festevam@denx.de>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Reviewed-by: Fabio Estevam <festevam@denx.de>
Link: https://lore.kernel.org/r/20230922192834.1695727-1-u.kleine-koenig@pengutronix.de
Signed-off-by: Lee Jones <lee@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit bbd2763ac8f4277a74195f6519c53242943503c3)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I80cb4137f6d9dfc54c9e2d41833c8f87b8764519
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83913
Tested-by: k2ci <kernel-bot@kylinos.cn>