crypto: hisilicon/qm - fix PF queue parameter issue
Mainline: 5831fc1fd4a578232fea708b82de0c666ed17153
From: v6.7-rc1
Severity: Low

[ Upstream commit 5831fc1fd4a578232fea708b82de0c666ed17153 ]

If the queue isolation feature is enabled, the number of queues
supported by the device changes. When PF is enabled using the
current default number of queues, the default number of queues may
be greater than the number supported by the device. As a result,
the PF fails to be bound to the driver.

After modification, if queue isolation feature is enabled, when
the default queue parameter is greater than the number supported
by the device, the number of enabled queues will be changed to
the number supported by the device, so that the PF and driver
can be properly bound.

Fixes: 8bbecfb402f7 ("crypto: hisilicon/qm - add queue isolation support for Kunpeng930")
Signed-off-by: Longfang Liu <liulongfang@huawei.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 4c79c7a4e7fbe63f4cbcd0785e2f3b6ca12d9eb0)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I90e6d748431b161ca108fce75aa861518f8997d7
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83860
Tested-by: k2ci <kernel-bot@kylinos.cn>