firmware: arm_ffa: Allow the FF-A drivers to use 32bit mode of messaging
Mainline: 2d698e8b4fd22374dac0a2d5150ab24d57a222ab
From: v6.7-rc1
Severity: Low

[ Upstream commit 2d698e8b4fd22374dac0a2d5150ab24d57a222ab ]

An FF-A ABI could support both the SMC32 and SMC64 conventions.
A callee that runs in the AArch64 execution state and implements such
an ABI must implement both SMC32 and SMC64 conventions of the ABI.

So the FF-A drivers will need the option to choose the mode irrespective
of FF-A version and the partition execution mode flag in the partition
information.

Let us remove the check on the FF-A version for allowing the selection
of 32bit mode of messaging. The driver will continue to set the 32-bit
mode if the partition execution mode flag specified that the partition
supports only 32-bit execution.

Fixes: 106b11b1ccd5 ("firmware: arm_ffa: Set up 32bit execution mode flag using partiion property")
Link: https://lore.kernel.org/r/20231005142823.278121-1-sudeep.holla@arm.com
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 38dd0c2905c716ae484281fcdc65d77aa8b0479b)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I63e186a984478d363fec1855a85e7e19684342b8
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83810
Tested-by: k2ci <kernel-bot@kylinos.cn>