media: cadence: csi2rx: Unregister v4l2 async notifier
Mainline: b2701715301a49b53d05c7d43f3fedc3b8743bfc
From: v6.7-rc1
Severity: Low

[ Upstream commit b2701715301a49b53d05c7d43f3fedc3b8743bfc ]

The notifier is added to the global notifier list when registered. When
the module is removed, the struct csi2rx_priv in which the notifier is
embedded, is destroyed. As a result the notifier list has a reference to
a notifier that no longer exists. This causes invalid memory accesses
when the list is iterated over. Similar for when the probe fails.
Unregister and clean up the notifier to avoid this.

Fixes: 1fc3b37f34f6 ("media: v4l: cadence: Add Cadence MIPI-CSI2 RX driver")

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Tested-by: Julien Massot <julien.massot@collabora.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Reviewed-by: Maxime Ripard <mripard@kernel.org>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit c1c88d6692e86492134a397a4eeb53095eb65fd4)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: Ie65e17ac1b3cff528a7f5d01d15d4bef04509b40
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/84022
Tested-by: k2ci <kernel-bot@kylinos.cn>