net: skb_find_text: Ignore patterns extending past 'to'
Mainline: c4eee56e14fe001e1cff54f0b438a5e2d0dd7454
From: v6.7-rc1
Severity: Low

[ Upstream commit c4eee56e14fe001e1cff54f0b438a5e2d0dd7454 ]

Assume that caller's 'to' offset really represents an upper boundary for
the pattern search, so patterns extending past this offset are to be
rejected.

The old behaviour also was kind of inconsistent when it comes to
fragmentation (or otherwise non-linear skbs): If the pattern started in
between 'to' and 'from' offsets but extended to the next fragment, it
was not found if 'to' offset was still within the current fragment.

Test the new behaviour in a kselftest using iptables' string match.

Suggested-by: Pablo Neira Ayuso <pablo@netfilter.org>
Fixes: f72b948dcbb8 ("[NET]: skb_find_text ignores to argument")
Signed-off-by: Phil Sutter <phil@nwl.cc>
Reviewed-by: Florian Westphal <fw@strlen.de>
Reviewed-by: Pablo Neira Ayuso <pablo@netfilter.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 7802db1dd578adcf5bef23ba3a3a9d396292dd2f)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I9faea7a11de51589a7c421a3395a9616d4b7b0b7
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83625
Tested-by: k2ci <kernel-bot@kylinos.cn>