ACPI: sysfs: Fix create_pnp_modalias() and create_of_modalias()
Mainline: 48cf49d31994ff97b33c4044e618560ec84d35fb
From: v6.7-rc1
Severity: Low

[ Upstream commit 48cf49d31994ff97b33c4044e618560ec84d35fb ]

snprintf() does not return negative values on error.

To know if the buffer was too small, the returned value needs to be
compared with the length of the passed buffer. If it is greater or
equal, the output has been truncated, so add checks for the truncation
to create_pnp_modalias() and create_of_modalias(). Also make them
return -ENOMEM in that case, as they already do that elsewhere.

Moreover, the remaining size of the buffer used by snprintf() needs to
be updated after the first write to avoid out-of-bounds access as
already done correctly in create_pnp_modalias(), but not in
create_of_modalias(), so change the latter accordingly.

Fixes: 8765c5ba1949 ("ACPI / scan: Rework modalias creation when "compatible" is present")
Signed-off-by: Christophe JAILLET <christophe.jaillet@wanadoo.fr>
[ rjw: Merge two patches into one, combine changelogs, add subject ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 60c874a7056ab025723f9e6ef1e629800bd29899)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I87721dfeb2da180e03ad783eb7104b53b8112d58
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/83651
Tested-by: k2ci <kernel-bot@kylinos.cn>