Fix termination state for idr_for_each_entry_ul()
Mainline: e8ae8ad479e2d037daa33756e5e72850a7bd37a9
From: v6.7-rc1
Severity: Low

[ Upstream commit e8ae8ad479e2d037daa33756e5e72850a7bd37a9 ]

The comment for idr_for_each_entry_ul() states

  after normal termination @entry is left with the value NULL

This is not correct in the case where UINT_MAX has an entry in the idr.
In that case @entry will be non-NULL after termination.
No current code depends on the documentation being correct, but to
save future code we should fix it.

Also fix idr_for_each_entry_continue_ul().  While this is not documented
as leaving @entry as NULL, the mellanox driver appears to depend on
it doing so.  So make that explicit in the documentation as well as in
the code.

Fixes: e33d2b74d805 ("idr: fix overflow case for idr_for_each_entry_ul()")
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Chris Mi <chrism@mellanox.com>
Cc: Cong Wang <xiyou.wangcong@gmail.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 4db74c184af9535db96c51714ca8f30c7ab768e5)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I272da1cf68211a269bdb4f545d965672c7cd4f30
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/84051
Tested-by: k2ci <kernel-bot@kylinos.cn>