cifs: do not reset chan_max if multichannel is not supported at mount
Mainline: 6e5e64c9477d58e73cb1a0e83eacad1f8df247cf
From: v6.7-rc1
Severity: Moderate

task: 186574

commit 6e5e64c9477d58e73cb1a0e83eacad1f8df247cf upstream.

If the mount command has specified multichannel as a mount option,
but multichannel is found to be unsupported by the server at the time
of mount, we set chan_max to 1. Which means that the user needs to
remount the share if the server starts supporting multichannel.

This change removes this reset. What it means is that if the user
specified multichannel or max_channels during mount, and at this
time, multichannel is not supported, but the server starts supporting
it at a later point, the client will be capable of scaling out the
number of channels.

Cc: stable@vger.kernel.org
Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(backported from commit 328004e6df5a9a8692e4ee04b3330b6b8931dba2)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: I14486a7ca56026e7ddc2c555731b9462c21b854d
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/85727
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>