KYLIN: arm64/config: enable CONFIG_ARM64_PTR_AUTH
Mainline: KYLIN-only
Severity: Low

Pointer authentication (part of the ARMv8.3 Extensions) provides
instructions for signing and authenticating pointers against secret
keys, which can be used to mitigate Return Oriented Programming (ROP)
and other attacks.

This option enables these instructions at EL0 (i.e. for userspace).
Choosing this option will cause the kernel to initialise secret keys
for each process at exec() time, with these keys being
context-switched along with the process.

The feature is detected at runtime. If the feature is not present in
hardware it will not be advertised to userspace/KVM guest nor will it
be enabled.

If the feature is present on the boot CPU but not on a late CPU, then
the late CPU will be parked. Also, if the boot CPU does not have
address auth and the late CPU has then the late CPU will still boot
but with the feature disabled. On such a system, this option should
not be selected.

K2CI-Arch: Arm64
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I81358a8045bde4c93c6fa9a9157f786c9a817e52
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/82317
Tested-by: k2ci <kernel-bot@kylinos.cn>