KYLIN: config: enable CONFIG_ACPI_PFRUT=m
Mainline: KYLIN-only
Severity: Low

This mechanism allows certain pieces of the platform firmware
to be updated on the fly while the system is running (runtime)
without the need to restart it, which is key in the cases when
the system needs to be available 100% of the time and it cannot
afford the downtime related to restarting it, or when the work
carried out by the system is particularly important, so it cannot
be interrupted, and it is not practical to wait until it is complete.

The existing firmware code can be modified (driver update) or
extended by adding new code to the firmware (code injection).

Besides, the telemetry driver allows user space to fetch telemetry
data from the firmware with the help of the Platform Firmware Runtime
Telemetry interface.

K2CI-Arch: All
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I480a924706fe4d05bc131967157abea26880fa0e
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/82319
Tested-by: k2ci <kernel-bot@kylinos.cn>