btrfs: make error messages more clear when getting a chunk map
Mainline: 7d410d5efe04e42a6cd959bfe6d59d559fdf8b25
From: v6.7-rc4
Severity: Moderate

task: 188294

commit 7d410d5efe04e42a6cd959bfe6d59d559fdf8b25 upstream.

When getting a chunk map, at btrfs_get_chunk_map(), we do some sanity
checks to verify we found a chunk map and that map found covers the
logical address the caller passed in. However the messages aren't very
clear in the sense that don't mention the issue is with a chunk map and
one of them prints the 'length' argument as if it were the end offset of
the requested range (while the in the string format we use %llu-%llu
which suggests a range, and the second %llu-%llu is actually a range for
the chunk map). So improve these two details in the error messages.

CC: stable@vger.kernel.org # 5.4+
Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Filipe Manana <fdmanana@suse.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(backported from commit 47ec6065fcd8bd0139676247689be341d15fdf6e)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: Ifdf622cf77ddcf97d4e6186020f4ca05d73c5429
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/87505
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>